+++ src/main/java/org/c4marathon/assignment/domain/auth/controller/AuthController.java
@@ -1,9 +1,7 @@
 package org.c4marathon.assignment.domain.auth.controller;
 
 import org.c4marathon.assignment.domain.auth.dto.request.SignUpRequest;
-import org.c4marathon.assignment.domain.consumer.service.ConsumerService;
-import org.c4marathon.assignment.domain.deliverycompany.service.DeliveryCompanyService;
-import org.c4marathon.assignment.domain.seller.service.SellerService;
+import org.c4marathon.assignment.domain.auth.service.AuthService;
 import org.c4marathon.assignment.global.constant.MemberType;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -21,17 +19,11 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/auth")
 public class AuthController {
 
-	private final ConsumerService consumerService;
-	private final SellerService sellerService;
-	private final DeliveryCompanyService deliveryCompanyService;
+	private final AuthService authService;
 
 	@ResponseStatus(HttpStatus.CREATED)
 	@PostMapping("/signup")
 	public void signup(@RequestBody @Valid SignUpRequest request, @RequestParam MemberType memberType) {
-		switch (memberType) {
-			case CONSUMER -> consumerService.signup(request);
-			case SELLER -> sellerService.signup(request);
-			case DELIVERY_COMPANY -> deliveryCompanyService.signup(request);
-		}
+		authService.signup(request, memberType);
 	}
 }
